# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Skill Swap API
  version: 1.0.0
  description: |
    HTTP contract implemented by the Sprint 1-2 backend.

    Authentication (FR-01) uses an opaque server-side session stored in the HttpOnly
    `skillswap_session` cookie (SameSite=Lax, Path=/). Browsers should call the API through
    the frontend origin (the Next.js `/api/v1/*` rewrite) with `credentials: "include"`.

    Every POST must come from the configured frontend origin (`Origin` header) and send
    `Content-Type: application/json` when it has a body. POST requests that need a session
    must also send the `X-CSRF-Token` header returned by `/auth/login` and `/auth/me`.
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: Local development
...
paths:
...
  /auth/verify-email:
    post:
      tags: [Auth]
      operationId: verifyEmail
      summary: Confirm the email address with the token from the verification link
      description: A token works once. Used, replaced, or expired tokens return `invalid_token`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [verified]
        "400":
          $ref: "#/components/responses/InvalidToken"
        "403":
          $ref: "#/components/responses/OriginForbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "503":
          $ref: "#/components/responses/DatabaseUnavailable"
  /auth/resend-verification:
    post:
      tags: [Auth]
      operationId: resendVerification
      summary: Send a new verification link to the signed-in user
      description: The new link invalidates previously sent verification links.
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "202":
          description: Email sent
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [sent]
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/CSRFOrOriginForbidden"
        "409":
          description: The email is already verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: email_already_verified
                  message: Email is already verified
        "503":
          description: Database (`database_unavailable`) or SMTP (`mail_unavailable`) is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      tags: [Auth]
      operationId: forgotPassword
      summary: Send a password reset link if an active account exists
      description: |
        Always answers 202 for a well-formed email so the response does not reveal whether an
        account exists. The link points to `{FRONTEND_ORIGIN}/reset-password?token=...`, is valid
        for 1 hour, and invalidates earlier reset links.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "202":
          description: Request accepted
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [accepted]
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "403":
          $ref: "#/components/responses/OriginForbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "503":
          $ref: "#/components/responses/DatabaseUnavailable"
  /auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Set a new password with the token from the reset link
      description: |
        On success every session of the user is revoked and a login lockout is lifted.
        The user then signs in with the new password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [password_reset]
        "400":
          $ref: "#/components/responses/InvalidToken"
        "403":
          $ref: "#/components/responses/OriginForbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "503":
          $ref: "#/components/responses/DatabaseUnavailable"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: skillswap_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: The `csrfToken` value from `/auth/login` or `/auth/me`.
      schema:
        type: string
  responses:
    Unauthenticated:
      description: The session cookie is missing, expired, or revoked
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: unauthenticated
              message: Sign in to continue
    CSRFOrOriginForbidden:
      description: Missing or wrong X-CSRF-Token (`csrf_invalid`) or foreign origin (`origin_forbidden`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: csrf_invalid
              message: CSRF token is missing or invalid
    InvalidJSON:
      description: The body is not a single JSON object with the documented fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: invalid_json
              message: Request body must be a valid JSON object
    UnsupportedMediaType:
      description: The body is not sent as application/json
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: unsupported_media_type
              message: Content-Type must be application/json
    ValidationFailed:
      description: One or more fields are invalid; `fields` maps each field name to a message
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: validation_failed
              message: Some fields are invalid
              fields:
                password: Password must contain at least 12 characters
    InvalidToken:
      description: The link token is malformed, already used, replaced, or expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: invalid_token
              message: The link is invalid, already used, or expired
    OriginForbidden:
      description: Request Origin does not match the configured frontend origin
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: origin_forbidden
              message: Origin is not allowed
    DatabaseUnavailable:
      description: PostgreSQL is unavailable or the required schema has not been applied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: database_unavailable
              message: Database is unavailable
  schemas:
...
    EmailRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
    TokenRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          description: The `token` query parameter from the email link
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 12
          description: 12+ characters, at most 72 UTF-8 bytes
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              description: Field-level messages, present only for validation_failed.
              additionalProperties:
                type: string